pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "council-engine"
version = "0.3.0"
description = "A structured deliberation protocol across diverse AI models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Amith Mathew", email = "amith@councilengine.dev" },
]
keywords = ["ai", "llm", "multi-model", "deliberation", "debate", "heterogeneous", "protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "google-genai>=1.64.0",
    "openai>=1.0.0",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
]

license-files = ["LICENSE"]

[project.urls]
Homepage = "https://councilengine.dev"
Repository = "https://github.com/amithmathew/council-engine"
Documentation = "https://councilengine.dev/architecture"

[tool.setuptools.packages.find]
include = ["council*"]

[project.scripts]
council = "council.cli:main"
